initialize most objects using logic classes, slightly modify addReview method in logic and dal so IDs are Longs, matching their type in their models

+++ instant_pot/src/main/java/com/web_services/instant_pot/dal/review/ReviewDAL.java
@@ -57,7 +57,7 @@ public class ReviewDAL {
 		return ReviewSet;
 	}
 	
-	public Review addReview(int customerID, int productID, String body) {	
+	public Review addReview(Long customerID, Long productID, String body) {	
 		SessionFactory sf = (SessionFactory) new Configuration().configure("hibernate.cfg.xml").buildSessionFactory();
 	    Session session = sf.openSession();
